Improve integration tests for FilterCommand and cleanup code

+++ src/main/java/seedu/address/model/person/WeeklyAttendanceList.java
@@ -38,7 +38,7 @@ public interface WeeklyAttendanceList extends Comparable<WeeklyAttendanceList> ,
      * Calculates the amount of absences
      * @return the number of absences
      */
-    public double calculateWeekAbsence();
+    public int calculateWeekAbsence();
 
     /**
      * @return attendence each week
